# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_beuphecan_remote-time-tracker_internal_dto.AcceptInvitationRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.BatchSyncRequest:
    properties:
      device_info:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SyncDeviceInfoItem'
      device_uuid:
        type: string
      organization_id:
        description: Default organization ID for all items
        type: integer
      screenshots:
        items:
          $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SyncScreenshotItem'
        type: array
      time_logs:
        items:
          $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SyncTimeLogItem'
        type: array
      workspace_id:
        description: Default workspace ID for all items
        type: integer
    required:
    - device_uuid
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.BatchSyncResponse:
    properties:
      device_info:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.DeviceInfoResponse'
      message:
        type: string
      screenshots_sync:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SyncResult'
      success:
        type: boolean
      synced_at:
        type: string
      time_logs_sync:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SyncResult'
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.DeviceInfoResponse:
    properties:
      app_version:
        type: string
      device_name:
        type: string
      device_uuid:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_seen_at:
        type: string
      os:
        type: string
      os_version:
        type: string
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.InvitationResponse:
    properties:
      accepted_at:
        type: string
      created_at:
        type: string
      email:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      invite_link:
        type: string
      invited_by:
        type: integer
      inviter:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.UserResponse'
      message:
        type: string
      org_role:
        type: string
      organization:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.OrganizationResponse'
      organization_id:
        type: integer
      status:
        type: string
      token:
        description: Only shown to creator
        type: string
      workspace:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.WorkspaceResponse'
      workspace_id:
        type: integer
      workspace_role:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.WorkspaceRoleResponse'
      workspace_role_id:
        type: integer
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.OrganizationMemberResponse:
    properties:
      id:
        type: integer
      invited_by:
        type: integer
      is_active:
        type: boolean
      joined_at:
        type: string
      role:
        type: string
      user:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.UserResponse'
      user_id:
        type: integer
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.OrganizationResponse:
    properties:
      allow_invite_link:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      invite_code:
        type: string
      is_active:
        type: boolean
      logo_url:
        type: string
      max_members:
        type: integer
      member_count:
        type: integer
      members:
        items:
          $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.OrganizationMemberResponse'
        type: array
      name:
        type: string
      owner:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.UserResponse'
      owner_id:
        type: integer
      share_invite_code:
        description: If true, all members can see invite code
        type: boolean
      slug:
        type: string
      updated_at:
        type: string
      workspace_count:
        type: integer
      workspaces:
        items:
          $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.WorkspaceResponse'
        type: array
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.SyncDeviceInfoItem:
    properties:
      app_version:
        type: string
      device_name:
        type: string
      device_uuid:
        type: string
      ip_address:
        type: string
      os:
        type: string
      os_version:
        type: string
    required:
    - device_uuid
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.SyncResult:
    properties:
      errors:
        items:
          type: string
        type: array
      failed:
        type: integer
      success:
        type: integer
      total:
        type: integer
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.SyncScreenshotItem:
    properties:
      base64_data:
        description: For file upload
        type: string
      captured_at:
        type: string
      checksum:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      is_encrypted:
        type: boolean
      local_id:
        type: string
      mime_type:
        type: string
      organization_id:
        description: Organization ID
        type: integer
      screen_number:
        type: integer
      task_id:
        description: 'Deprecated: Use TaskLocalID instead'
        type: integer
      task_local_id:
        description: UUID from Electron - primary task identifier
        type: string
      time_log_id:
        description: Server-side TimeLog ID (deprecated, use TimeLogLocalID)
        type: integer
      time_log_local_id:
        description: Local ID of TimeLog from Electron
        type: string
      workspace_id:
        description: Workspace ID the screenshot belongs to
        type: integer
    required:
    - captured_at
    - file_name
    - local_id
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.SyncTimeLogItem:
    properties:
      duration:
        type: integer
      end_time:
        type: string
      local_id:
        type: string
      notes:
        type: string
      organization_id:
        description: Organization ID
        type: integer
      paused_at:
        type: string
      paused_total:
        type: integer
      resumed_at:
        type: string
      start_time:
        type: string
      status:
        type: string
      task_id:
        description: 'Deprecated: Use TaskLocalID instead'
        type: integer
      task_local_id:
        description: UUID from Electron - primary task identifier
        type: string
      task_title:
        description: Task title when stopped
        type: string
      workspace_id:
        description: Workspace ID the time log belongs to
        type: integer
    required:
    - local_id
    - start_time
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      role:
        type: string
    type: object
...
  github_com_beuphecan_remote-time-tracker_internal_dto.WorkspaceMemberResponse:
    properties:
      added_by:
        type: integer
      can_manage_tasks:
        type: boolean
      can_view_reports:
        type: boolean
      id:
        type: integer
      is_active:
        type: boolean
      is_admin:
        type: boolean
      joined_at:
        type: string
      role_name:
        type: string
      user:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.UserResponse'
      user_id:
        type: integer
      workspace_role:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.WorkspaceRoleResponse'
      workspace_role_id:
        type: integer
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.WorkspaceResponse:
    properties:
      admin:
        $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.UserResponse'
      admin_id:
        type: integer
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      end_date:
        type: string
      hourly_rate:
        type: number
      icon:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_billable:
        type: boolean
      member_count:
        type: integer
      members:
        items:
          $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.WorkspaceMemberResponse'
        type: array
      name:
        type: string
      organization_id:
        type: integer
      slug:
        type: string
      start_date:
        type: string
      task_count:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_beuphecan_remote-time-tracker_internal_dto.WorkspaceRoleResponse:
    properties:
      color:
        type: string
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      organization_id:
        type: integer
      permissions:
        type: string
      sort_order:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@remotetimetracker.com
    name: API Support
  description: API documentation for Remote Time Tracker - A comprehensive time tracking
    application with screenshot capture, task management, organization/workspace management,
    and synchronization features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Remote Time Tracker API
  version: "1.0"
paths:
...
  /invitations/{id}:
    delete:
      description: Revoke a pending invitation
      parameters:
      - description: Invitation ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Invitation revoked
        "400":
          description: Invalid invitation ID
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Revoke invitation
      tags:
      - invitations
  /invitations/{token}:
    get:
      description: Get invitation details by token (before accepting)
      parameters:
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invitation details
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.InvitationResponse'
        "400":
          description: Token required
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "404":
          description: Invitation not found or expired
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      summary: Get invitation by token
      tags:
      - invitations
  /invitations/accept:
    post:
      consumes:
      - application/json
      description: Accept an invitation with token provided in request body (for public
        registration flow)
      parameters:
      - description: Invitation token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.AcceptInvitationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Invitation accepted
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.OrganizationMemberResponse'
        "400":
          description: Invalid request or token
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      summary: Accept invitation by token in body
      tags:
      - invitations
  /invitations/accept/{token}:
    post:
      description: Accept an invitation by token (requires authentication)
      parameters:
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Invitation accepted
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.OrganizationMemberResponse'
        "400":
          description: Invalid token or already accepted
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Accept invitation
      tags:
      - invitations
  /invitations/my:
    get:
      description: Get all pending invitations for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Pending invitations
          schema:
            items:
              $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.InvitationResponse'
            type: array
        "400":
          description: User email not found
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my pending invitations
      tags:
      - invitations
...
  /sync/batch:
    post:
      consumes:
      - application/json
      description: Synchronize time logs, screenshots, and device info from Electron
        desktop app. Supports offline-first sync with conflict resolution.
      parameters:
      - description: Batch sync request containing time logs, screenshots, and device
          info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.BatchSyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Batch sync completed
          schema:
            allOf:
            - $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.BatchSyncResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
        "500":
          description: Sync failed
          schema:
            $ref: '#/definitions/github_com_beuphecan_remote-time-tracker_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch sync data from desktop app
      tags:
      - sync
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...